int→int64 problems: floats in the corner added

+++ go/kernighan_donovan_2016/03_04_floats_SVG_server/svgServer.go
@@ -106,8 +106,8 @@ func corner(i, j int, f func(float64, float64) float64) (float64, float64, float
     z := f(x, y)                 // HERE you can put eggbox or saddle functions
 
     // project (x, y, z) isometrically into 2d svg canvas (sx, sy)
-    sx := we / 2.0 + (x - y) * cos30 * xyscale
-    sy := h / 2 + (x + y) * sin30 * xyscale - z * zscale
+    sx := float64(we) / 2 + (x - y) * cos30 * float64(xyscale)
+    sy := float64(h) / 2 + (x + y) * sin30 * float64(xyscale) - z * float64(zscale)
     return sx, sy, z
 }
